backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
qdrant-client==1.7.3
playwright==1.41.2
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3
sentence-transformers==2.3.1
torch==2.2.0
pillow==10.2.0
faster-whisper==0.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
